refactor(item): move message methods to partial, convert text params to ReadOnlySpan<char>
