+++ Projects/Server/Items/Item.cs
@@ -3312,68 +3312,6 @@ public partial class Item : IHued, IComparable<Item>, ISpawnable, IObjectPropert
         Delete();
     }
 
-    public void PublicOverheadMessage(MessageType type, int hue, bool ascii, string text)
-    {
-        if (m_Map == null)
-        {
-            return;
-        }
-
-        var worldLoc = GetWorldLocation();
-
-        var buffer = stackalloc byte[OutgoingMessagePackets.GetMaxMessageLength(text)].InitializePacket();
-
-        foreach (var state in m_Map.GetClientsInRange(worldLoc, GetMaxUpdateRange()))
-        {
-            var m = state.Mobile;
-
-            if (m.CanSee(this) && m.InRange(worldLoc, GetUpdateRange(m)))
-            {
-                var length = OutgoingMessagePackets.CreateMessage(
-                    buffer, Serial, m_ItemID, type, hue, 3, ascii, "ENU", Name, text
-                );
-
-                if (length != buffer.Length)
-                {
-                    buffer = buffer[..length]; // Adjust to the actual size
-                }
-
-                state.Send(buffer);
-            }
-        }
-    }
-
-    public void PublicOverheadMessage(MessageType type, int hue, int number, string args = "")
-    {
-        if (m_Map == null)
-        {
-            return;
-        }
-
-        var worldLoc = GetWorldLocation();
-
-        var buffer = stackalloc byte[OutgoingMessagePackets.GetMaxMessageLocalizedLength(args)].InitializePacket();
-
-        foreach (var state in m_Map.GetClientsInRange(worldLoc, GetMaxUpdateRange()))
-        {
-            var m = state.Mobile;
-
-            if (m.CanSee(this) && m.InRange(worldLoc, GetUpdateRange(m)))
-            {
-                var length = OutgoingMessagePackets.CreateMessageLocalized(
-                    buffer, Serial, m_ItemID, type, hue, 3, number, Name, args
-                );
-
-                if (length != buffer.Length)
-                {
-                    buffer = buffer[..length]; // Adjust to the actual size
-                }
-
-                state.Send(buffer);
-            }
-        }
-    }
-
     public virtual void OnAfterDelete()
     {
     }
@@ -3895,37 +3833,6 @@ public partial class Item : IHued, IComparable<Item>, ISpawnable, IObjectPropert
             m_Location.m_X,
             m_Location.m_Y,
             m_Location.m_Z + ItemData.CalcHeight
-        );
-    }
-
-    public void SendLocalizedMessageTo(Mobile to, int number, string args = "")
-    {
-        if (Deleted || !to.CanSee(this))
-        {
-            return;
-        }
-
-        to.NetState.SendMessageLocalized(Serial, ItemID, MessageType.Regular, 0x3B2, 3, number, "", args);
-    }
-
-    public void SendLocalizedMessageTo(Mobile to, int number, AffixType affixType, string affix, string args)
-    {
-        if (Deleted || !to.CanSee(this))
-        {
-            return;
-        }
-
-        to.NetState.SendMessageLocalizedAffix(
-            Serial,
-            ItemID,
-            MessageType.Regular,
-            0x3B2,
-            3,
-            number,
-            "",
-            affixType,
-            affix,
-            args
         );
     }
